+++ src/backend/postgres.rs
@@ -1,195 +1,249 @@
-use crate::models::{FeatureFlag, GroupSet};
-use crate::schema::fun_with_flags_toggles;
-use diesel::prelude::*;
-use diesel::sql_types::Bool;
+use crate::models::{FeatureFlag, GroupSet, RawFeatureFlag};
 
-pub type DB = diesel::pg::Pg;
-pub type DBConnection = PgConnection;
-pub type SetOutput = Result<Vec<FeatureFlag>, diesel::result::Error>;
-pub type GetOutput = Result<FeatureFlag, diesel::result::Error>;
+use postgres::types::ToSql;
+use postgres::NoTls;
+use postgres::Row;
+
+use r2d2_postgres::PostgresConnectionManager;
+
+pub type DB = ();
+pub type DBConnection = Connection;
+pub type SetOutput = Result<FeatureFlag, Error>;
+pub type GetOutput = Result<FeatureFlag, Error>;
+
+#[derive(Debug)]
+pub struct Error(String);
+
+impl std::fmt::Display for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl From<r2d2::Error> for Error {
+    fn from(e: r2d2::Error) -> Self {
+        Error(e.to_string())
+    }
+}
+
+impl From<postgres::Error> for Error {
+    fn from(e: postgres::Error) -> Self {
+        Error(e.to_string())
+    }
+}
+
+pub struct Connection {
+    pub pool: r2d2::Pool<PostgresConnectionManager<NoTls>>,
+}
+
+impl Connection {
+    pub fn establish(url: &str) -> Result<Connection, ()> {
+        let manager = PostgresConnectionManager::new(url.parse().unwrap(), NoTls);
+        let pool = r2d2::Pool::new(manager).unwrap();
+        Ok(Connection { pool })
+    }
+}
 
 pub struct Backend {}
 
 impl Backend {
-    pub fn set(
-        conn: &DBConnection,
-        flag: FeatureFlag,
-    ) -> Result<Vec<FeatureFlag>, diesel::result::Error> {
-        use crate::schema::fun_with_flags_toggles::dsl::*;
+    pub fn set(conn: &DBConnection, flag: FeatureFlag) -> SetOutput {
+        let mut conn = Self::create_conn(conn)?;
 
-        match flag {
-            FeatureFlag::Percentage { .. } | FeatureFlag::Time { .. } => {
-                let insertable = flag.to_insertable();
+        use FeatureFlag::*;
 
-                let filter = fun_with_flags_toggles.filter(
-                    flag_name
-                        .eq(insertable.flag_name)
-                        .and(gate_type.eq(insertable.gate_type)),
-                );
-
-                let fetch_result = if *insertable.enabled {
-                    diesel::update(filter)
-                        .set((target.eq(insertable.target), enabled.eq(insertable.enabled)))
-                        .get_results::<FeatureFlag>(conn)
+        let insertable = flag.to_row();
+        let db_result = match flag {
+            Percentage { .. } | Time { .. } => {
+                let result = if insertable.enabled {
+                    let update_enable = r#"UPDATE "fun_with_flags_toggles" 
+                SET "target" = $1, "enabled" = $2 
+                WHERE "fun_with_flags_toggles"."flag_name" = $3 AND "fun_with_flags_toggles"."gate_type" = $4 
+                RETURNING "fun_with_flags_toggles"."id", "fun_with_flags_toggles"."flag_name", "fun_with_flags_toggles"."gate_type", "fun_with_flags_toggles"."target", "fun_with_flags_toggles"."enabled""#;
+                    let arg_enable1: Vec<&(dyn ToSql + Sync)> = vec![
+                        &insertable.target,
+                        &insertable.enabled,
+                        &insertable.flag_name,
+                        &insertable.gate_type,
+                    ];
+                    conn.query_opt(update_enable, &arg_enable1)?
                 } else {
-                    diesel::update(filter)
-                        .set(enabled.eq(insertable.enabled))
-                        .get_results::<FeatureFlag>(conn)
-                }?;
+                    let update_disable = r#"UPDATE "fun_with_flags_toggles" 
+                SET "enabled" = $1 
+                WHERE "fun_with_flags_toggles"."flag_name" = $2 AND "fun_with_flags_toggles"."gate_type" = $3 
+                RETURNING "fun_with_flags_toggles"."id", "fun_with_flags_toggles"."flag_name", "fun_with_flags_toggles"."gate_type", "fun_with_flags_toggles"."target", "fun_with_flags_toggles"."enabled""#;
+                    let arg_disable1: Vec<&(dyn ToSql + Sync)> = vec![
+                        &insertable.enabled,
+                        &insertable.flag_name,
+                        &insertable.gate_type,
+                    ];
+                    conn.query_opt(update_disable, &arg_disable1)?
+                };
 
-                if fetch_result.is_empty() {
-                    diesel::insert_into(fun_with_flags_toggles)
-                        .values(flag.to_insertable())
-                        .get_results::<FeatureFlag>(conn)
+                if result.is_none() {
+                    let insert = r#"INSERT INTO "fun_with_flags_toggles" ("flag_name", "gate_type", "target", "enabled") 
+                VALUES ($1, $2, $3, $4) 
+                RETURNING "fun_with_flags_toggles"."id", "fun_with_flags_toggles"."flag_name", "fun_with_flags_toggles"."gate_type", "fun_with_flags_toggles"."target", "fun_with_flags_toggles"."enabled""#;
+                    let arg2: Vec<&(dyn ToSql + Sync)> = vec![
+                        &insertable.flag_name,
+                        &insertable.gate_type,
+                        &insertable.target,
+                        &insertable.enabled,
+                    ];
+                    conn.query_one(insert, &arg2)?
                 } else {
-                    Ok(fetch_result)
+                    result.unwrap()
                 }
             }
-            _ => diesel::insert_into(fun_with_flags_toggles)
-                .values(flag.to_insertable())
-                .on_conflict((flag_name, gate_type, target))
-                .do_update()
-                .set(enabled.eq(flag.enabled()))
-                .get_results::<FeatureFlag>(conn),
-        }
+            _ => {
+                let insert = r#"INSERT INTO "fun_with_flags_toggles" ("flag_name", "gate_type", "target", "enabled") 
+            VALUES ($1, $2, $3, $4) ON CONFLICT ("flag_name", "gate_type", "target") 
+            DO UPDATE SET "enabled" = $5 
+            RETURNING "fun_with_flags_toggles"."id", "fun_with_flags_toggles"."flag_name", "fun_with_flags_toggles"."gate_type", "fun_with_flags_toggles"."target", "fun_with_flags_toggles"."enabled""#;
+                let args: Vec<&(dyn ToSql + Sync)> = vec![
+                    &insertable.flag_name,
+                    &insertable.gate_type,
+                    &insertable.target,
+                    &insertable.enabled,
+                    &insertable.enabled,
+                ];
+                conn.query_one(insert, &args)?
+            }
+        };
+
+        Ok(FeatureFlag::from_row(db_result))
     }
 
-    pub fn get(
-        conn: &DBConnection,
-        flag: FeatureFlag,
-    ) -> Result<FeatureFlag, diesel::result::Error> {
-        use crate::schema::fun_with_flags_toggles::dsl::*;
+    pub fn get(conn: &DBConnection, flag: FeatureFlag) -> GetOutput {
+        let mut conn = Self::create_conn(conn).unwrap();
 
-        fun_with_flags_toggles
-            .filter(flag.to_filter())
-            .first::<FeatureFlag>(conn)
+        use FeatureFlag::*;
+
+        let insertable = flag.to_row();
+        let db_result = match flag {
+            Time { .. } | Percentage { .. } | Group { .. } => {
+                let query = r#"SELECT "fun_with_flags_toggles"."id", "fun_with_flags_toggles"."flag_name", "fun_with_flags_toggles"."gate_type", "fun_with_flags_toggles"."target", "fun_with_flags_toggles"."enabled" 
+                FROM "fun_with_flags_toggles" WHERE "fun_with_flags_toggles"."flag_name" = $1 AND "fun_with_flags_toggles"."gate_type" = $2"#;
+                let arguments: Vec<&(dyn ToSql + Sync)> =
+                    vec![&insertable.flag_name, &insertable.gate_type];
+                conn.query_one(query, &arguments)?
+            }
+
+            _ => {
+                let query = r#"SELECT "fun_with_flags_toggles"."id", "fun_with_flags_toggles"."flag_name", "fun_with_flags_toggles"."gate_type", "fun_with_flags_toggles"."target", "fun_with_flags_toggles"."enabled" 
+                FROM "fun_with_flags_toggles" WHERE "fun_with_flags_toggles"."flag_name" = $1 AND "fun_with_flags_toggles"."gate_type" = $2 AND "fun_with_flags_toggles"."target" = $3"#;
+                let arguments: Vec<&(dyn ToSql + Sync)> = vec![
+                    &insertable.flag_name,
+                    &insertable.gate_type,
+                    &insertable.target,
+                ];
+                conn.query_one(query, &arguments)?
+            }
+        };
+
+        Ok(FeatureFlag::from_row(db_result))
     }
 
     pub fn backend_name() -> &'static str {
         "postgres"
     }
-}
 
-impl Queryable<fun_with_flags_toggles::SqlType, DB> for FeatureFlag {
-    type Row = (i64, String, String, String, bool);
-
-    fn build(row: Self::Row) -> Self {
-        match row.2.as_ref() {
-            "boolean" => FeatureFlag::Boolean {
-                name: row.1,
-                enabled: row.4,
-            },
-            "actor" => FeatureFlag::Actor {
-                name: row.1,
-                target: row.3,
-                enabled: row.4,
-            },
-            "group" => FeatureFlag::Group {
-                name: row.1,
-                target: GroupSet::new(row.3),
-                enabled: row.4,
-            },
-            "percentage" if row.3.starts_with("time/") => FeatureFlag::Time {
-                name: row.1,
-                target: parse_float(row.3, "time/"),
-                enabled: row.4,
-            },
-            "percentage" if row.3.starts_with("actors/") => FeatureFlag::Percentage {
-                name: row.1,
-                target: parse_float(row.3, "actors/"),
-                enabled: row.4,
-            },
-            _ => panic!("this gate is not supported"),
-        }
+    pub fn create_conn(
+        pool: &DBConnection,
+    ) -> Result<r2d2::PooledConnection<PostgresConnectionManager<NoTls>>, r2d2::Error> {
+        let pool = pool.pool.clone();
+        pool.get()
     }
 }
 
 impl FeatureFlag {
-    pub fn to_insertable<'a>(&'a self) -> NewFeatureFlag<'a> {
+    pub fn to_row(&self) -> RawFeatureFlag {
         use FeatureFlag::*;
 
         match self {
-            Boolean { name, enabled } => NewFeatureFlag {
-                flag_name: name,
-                gate_type: "boolean",
+            Boolean { name, enabled } => RawFeatureFlag {
+                flag_name: name.to_string(),
+                gate_type: "boolean".to_string(),
                 target: "_fwf_none".to_string(),
-                enabled: enabled,
+                enabled: *enabled,
             },
             Actor {
                 name,
                 target,
                 enabled,
-            } => NewFeatureFlag {
-                flag_name: name,
-                gate_type: "actor",
+            } => RawFeatureFlag {
+                flag_name: name.to_string(),
+                gate_type: "actor".to_string(),
                 target: target.to_string(),
-                enabled: enabled,
+                enabled: *enabled,
             },
             Group {
                 name,
                 target,
                 enabled,
-            } => NewFeatureFlag {
-                flag_name: name,
-                gate_type: "group",
+            } => RawFeatureFlag {
+                flag_name: name.to_string(),
+                gate_type: "group".to_string(),
                 target: target.to_optional_string().unwrap_or("").to_string(),
-                enabled: enabled,
+                enabled: *enabled,
             },
             Time {
                 name,
                 target,
                 enabled,
-            } => NewFeatureFlag {
-                flag_name: name,
-                gate_type: "percentage",
+            } => RawFeatureFlag {
+                flag_name: name.to_string(),
+                gate_type: "percentage".to_string(),
                 target: format!("time/{}", target),
-                enabled: enabled,
+                enabled: *enabled,
             },
             Percentage {
                 name,
                 target,
                 enabled,
-            } => NewFeatureFlag {
-                flag_name: name,
-                gate_type: "percentage",
+            } => RawFeatureFlag {
+                flag_name: name.to_string(),
+                gate_type: "percentage".to_string(),
                 target: format!("actors/{}", target),
-                enabled: enabled,
+                enabled: *enabled,
             },
         }
     }
 
-    // pub fn to_filter<'a>(&'a self) -> (&'a str, &'a str, String) {
-    pub fn to_filter<'a>(
-        &'a self,
-    ) -> Box<dyn BoxableExpression<fun_with_flags_toggles::table, DB, SqlType = Bool> + 'a> {
-        use crate::schema::fun_with_flags_toggles::dsl::*;
-        use FeatureFlag::*;
+    pub fn from_row(row: Row) -> FeatureFlag {
+        let flag_name: String = row.get("flag_name");
+        let gate_type: String = row.get("gate_type");
+        let target: String = row.get("target");
+        let enabled: bool = row.get("enabled");
 
-        let insertable = self.to_insertable();
-        match self {
-            Time { .. } | Percentage { .. } | Group { .. } => Box::new(
-                flag_name
-                    .eq(insertable.flag_name)
-                    .and(gate_type.eq(insertable.gate_type)),
-            ),
-
-            _ => Box::new(
-                flag_name
-                    .eq(insertable.flag_name)
-                    .and(gate_type.eq(insertable.gate_type))
-                    .and(target.eq(insertable.target)),
-            ),
+        match gate_type.as_ref() {
+            "boolean" => FeatureFlag::Boolean {
+                name: flag_name,
+                enabled,
+            },
+            "actor" => FeatureFlag::Actor {
+                name: flag_name,
+                target,
+                enabled,
+            },
+            "group" => FeatureFlag::Group {
+                name: flag_name,
+                target: GroupSet::new(target),
+                enabled,
+            },
+            "percentage" if target.starts_with("time/") => FeatureFlag::Time {
+                name: flag_name,
+                target: parse_float(target, "time/"),
+                enabled,
+            },
+            "percentage" if target.starts_with("actors/") => FeatureFlag::Percentage {
+                name: flag_name,
+                target: parse_float(target, "actors/"),
+                enabled,
+            },
+            _ => panic!("this gate is not supported"),
         }
     }
-}
-
-#[derive(Insertable, AsChangeset)]
-#[table_name = "fun_with_flags_toggles"]
-pub struct NewFeatureFlag<'a> {
-    pub flag_name: &'a str,
-    pub gate_type: &'a str,
-    pub target: String,
-    pub enabled: &'a bool,
 }
 
 fn parse_float(mut value: String, prefix: &'static str) -> f64 {
