--- src/schema.rs
@@ -1,9 +1,0 @@
-table! {
-    fun_with_flags_toggles (id) {
-        id -> Int8,
-        flag_name -> Varchar,
-        gate_type -> Varchar,
-        target -> Varchar,
-        enabled -> Bool,
-    }
-}
